New Law Sets And Minor Fixes on AI (#1686)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Added 4 new law boards for AI:

1- NT Agressive
2- Quarantine
3- CCTV
4- TYRANT

Now the AI ​​can see through Borg and has become a target for the Ion
Storm

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Ichai
- add: Added 4 law boards to the game: NT Agressive, Quarantine, CCTV
and TYRANT to the AI
- tweak: AI can see through borgs
- tweak: AI is now affected by ion storm

---------

Signed-off-by: Ichaie <167008606+Ichaie@users.noreply.github.com>