The Shotgun Update Part 2 (#2004)
This PR makes use of the Shotgun Refactor by adding a new type of
special ammunition called "Shrapnel" ammo, as well as a variety of new
shotgun shell types. There's also new 40mm shotshell ammo types, which
let you turn a china lake into a 40mm shotgun of DOOM.

Shrapnel ammo essentially makes "Any gun into a shotgun", splitting the
damage of the original base ammo across multiple projectiles in a cone.
It has a small damage multiplier as a tradeoff, such that they deal
increased damage at incredibly close range. But realistically, you want
this ammo for the hilarious fun that is throwing 4x as many bullets down
a hallway.

Shrapnel ammo can be bought in the Syndicate Uplink, alongside the other
varieties of exotic special ammo, and can even be purchased in a
Shrapnel Ammo Bundle for nukie teams. To better encourage players to not
feel like they're being punished for taking special ammo, most of the
traitor weapons no longer spawn with ammo, and are in turn discounted by
the cost of the ammo they no longer spawn with. The C20r for instance
has gone all the way down to 50tc.

<details><summary><h1>Media</h1></summary>
<p>

8mb video isn't working right now.

</p>
</details>

:cl:
- add: Added "Shrapnel Ammo" to most weapon types as a new special
ammunition. Shrapnel ammo essentially turns any gun into a baby shotgun.
This includes guns like the L6 Saw.
- add: Added Shrapnel Ammo and specialty ammo bundles to the Syndicate
Uplink.
- tweak: To encourage people to buy different kinds of ammo, certain
guns in the Syndicate Uplink no longer spawn with ammo, and are in
return discounted by the cost of the ammo that was removed. Have fun
buying C20 magazines with incendiary ammo.
- add: Added Shrapnel Munitions to Arsenal Research T2. Once researched,
shrapnel ammo can be manufactured at a security techfab.